LayoutTests/accessibility/element-reflection-ariacontrols-expected.txt

Checks that element reflection is exposed to the a11y tree for 'ariaControlsElements'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axTab1.ariaControlsElementAtIndex(0).isEqual(axPanel1) is true
PASS axTab2.ariaControlsElementAtIndex(0).isEqual(axPanel2) is true
PASS axTab2.ariaControlsElementAtIndex(0).isEqual(axPanel1) is true
PASS axTab3.ariaControlsElementAtIndex(0).isEqual(axPanel3) is true
PASS axTab2.ariaControlsElementAtIndex(0).isEqual(axPanel1) is true
PASS axTab2.ariaControlsElementAtIndex(1).isEqual(axPanel2) is true
PASS axTab2.ariaControlsElementAtIndex(2).isEqual(axPanel3) is true
PASS axTab4.ariaControlsElementAtIndex(0).isEqual(axPanel4) is true
PASS axTab5.ariaControlsElementAtIndex(0).isEqual(axPanel5) is true
PASS successfullyParsed is true

TEST COMPLETE
First panel
Tab 1
Second panel
Tab 2
Third panel
Fourth panel
Tab 4
Fifth panel
